Sat, Jul 1, 2023, 4:20 PM +01:00

+++ professor/check.py
@@ -9,7 +9,7 @@ def main():
 
 def check(s):
     try:
-        t = int(s)
+        t = float(s)
         return t
     except ValueError:
         pass
